refactor and fix problems and change app icon

+++ lib/features/payment/data/data_source/payment_manager.dart
@@ -5,7 +5,7 @@ import 'package:flutter_stripe/flutter_stripe.dart';
 import 'package:restaurant/core/dependency_injection/service_locator.dart';
 
 class PaymentManager {
-  static Future<bool> makePayment(int amount, String currency) async {
+  static Future<bool> makePayment(double amount, String currency) async {
     try {
       String clientSecret = await _getClientSecret(
         (amount * 100).toString(),
